refactor(compliance): split data_collector.py (1559 LOC) into collectors/ package
Campaign B B2: the 1559-LOC compliance data_collector monolith was split
into per-domain modules under `plugins/shipwright-compliance/scripts/lib/
collectors/`. The original file becomes a 41-LOC re-export shim that
preserves the pre-split public + private import surface (every dataclass,
every collect_*, every probed helper). Behavior is bit-preserving: all 5
compliance MDs (rtm/test-evidence/change-history/sbom/dashboard)
regenerate byte-identical against the pre-split baseline.

Split layout (all <= 300 LOC):
* collectors/_types.py        — 12 public dataclasses (ComplianceData et al)
* collectors/_common.py       — CONFIG_FILES + collect_configs
* collectors/dashboard.py     — splits + sections + req->section mapper
* collectors/rtm.py           — decisions + requirements + ER states +
                                  req->event mapper
* collectors/test_evidence.py — test results + test files + known failures
* collectors/change_history.py — git log + event log (worktree-aware)
* collectors/sbom.py          — dependencies + per-workspace undeclared
* collectors/_npm_license.py  — npm license resolver (lockfile-first)
* collectors/_python_license.py — python license resolver (per-manifest
                                    .venv dist-info METADATA)
* collectors/__init__.py      — collect_all + re-exports

The B8 cross-plugin contract (`shared/contracts/compliance.py`) still
resolves `collect_all` + dataclasses through the shim unchanged. Two
integration tests that loaded data_collector via
`spec_from_file_location` (which fails on relative imports) were
updated to import via the package; the test_data_collector_shim.py
adds explicit coverage that the legacy shim surface keeps resolving.

Bloat baseline: the `data_collector.py` entry (1559 LOC, grandfathered)
is removed in this same commit — the new 41-LOC shim plus the per-domain
modules are all under their respective limits. No new grandfathered entry
is added.

External LLM review (OpenRouter; openai + gemini): completed.
Findings dispositioned in surface_verification.json + the iterate run
dir. Test-suite update for `tests/test_data_collector.py` was rejected
(campaign hard guard explicitly excludes that grandfathered file from
Campaign B refactor scope; the new test_data_collector_shim.py
explicitly pins the back-compat surface instead).

Empirical verification:
* 534 compliance plugin tests pass (vs. 80 in data_collector alone +
  the new 6 in test_data_collector_shim.py)
* 136 integration tests pass (incl. test_shared_contracts_consumers*
  and test_fr_table_drift_protection)
* MD round-trip probe: 5/5 byte-identical to pre-split baseline
