+++ plugins/shipwright-compliance/scripts/lib/collectors/_common.py
@@ -1,0 +1,33 @@
+"""Shared runtime helpers for compliance collectors.
+
+Hosts the small cross-collector utilities (config-file lookup + read).
+Dataclass definitions live in ``_types.py``; keeping helpers separate
+keeps both files well under the 300-LOC budget.
+
+Iterate Campaign B (B2): split out of ``data_collector.py``.
+"""
+
+from __future__ import annotations
+
+import json
+from pathlib import Path
+
+
+CONFIG_FILES = {
+    "run": "shipwright_run_config.json",
+    "project": "shipwright_project_config.json",
+    "plan": "shipwright_plan_config.json",
+    "build": "shipwright_build_config.json",
+}
+
+
+def collect_configs(project_root: Path) -> dict[str, dict]:
+    """Read all shipwright config files. Returns empty dicts for missing files."""
+    configs: dict[str, dict] = {}
+    for key, filename in CONFIG_FILES.items():
+        path = project_root / filename
+        if path.exists():
+            configs[key] = json.loads(path.read_text(encoding="utf-8"))
+        else:
+            configs[key] = {}
+    return configs
